seperate all rentals  methods into one Class and gather everything in userinterface class
