+++ main.rb
@@ -1,3 +1,5 @@
+#!/usr/bin/env ruby
+
 require './person'
 require './student'
 require './teacher'
@@ -7,66 +9,37 @@ require './rental'
 require './corrector'
 require './BookUi.rb'
 require './PersonUi.rb'
+require './RentalUi.rb'
+
 class UserInterface
   def initialize
     @people = PersonUserInterface.new
-    @rentals = []
     @book = BookUserInterface.new
+    @rental = RentalUserInterface.new
   end
 
-  def create_rental
-    if @books.count.zero?
-      puts 'Please try to create book first and try again !'
-    else
-      puts 'Select a book from the following list by number'
-      @books.each_with_index do |book, index|
-        puts "#{index})Title: #{book.title}, Author: #{book.author}"
-      end
-      book_num = gets.chomp
-    end
-    if @people.count.zero?
-      puts 'Please try to create PERSON first and try again !'
-    else
-      puts 'Select a person from the following list by number (not id)'
-      @people.each_with_index do |person, index|
-        puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-      end
-      person_num = gets.chomp
-    end
-
-    puts 'Date: '
-    date = gets.chomp
-
-    @rentals << Rental.new(@books[book_num.to_i], @people[person_num.to_i], date)
-    puts 'Rental created successfully'
-    puts ''
-  end
-
-  def list_all_rental_by_id
-    print 'ID of person: '
-    id = gets.chomp
-
-    filtred_rentals = @rentals.select { |rental| rental.person.id == id.to_i }
-    filtred_rentals.each do |rental|
-      puts "Date: #{rental.date}, Book: #{rental.book.title} by #{rental.book.author}"
-    end
-  end
-
-  def run
+  def start
     puts 'Welcome to Shcool library App!'
     puts ' '
-    loop do
-      puts 'Please choose an option by entring a number: '
-      puts '1 - List all books'
-      puts '2 - List all people'
-      puts '3 - Create a person'
-      puts '4 - Create a book'
-      puts '5 - Create a rental'
-      puts '6 - List all rentals for given person id'
-      puts '7 - Exit'
+    self.command_executor
+  end
 
+  private
+  def command_instruction
+    puts 'Please choose an option by entring a number: '
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Create a person'
+    puts '4 - Create a book'
+    puts '5 - Create a rental'
+    puts '6 - List all rentals for given person id'
+    puts '7 - Exit'
+  end
+
+  def command_executor
+      loop do
+      self.command_instruction
       answer = gets.chomp
-
       case answer
       when '1'
         @book.list_all_books
@@ -77,9 +50,9 @@ class UserInterface
       when '4'
         @book.create_book
       when '5'
-        create_rental
+        @rental.create_rental(@book.books,@people.people)
       when '6'
-        list_all_rental_by_id
+        @rental.list_all_rental_by_id
       when '7'
         break
       end
@@ -88,8 +61,8 @@ class UserInterface
 end
 
 def main
-  app = App.new
-  app.run
+  app = UserInterface.new
+  app.start
 end
 
 main
